refactor(schemas): delete services/lib/unified-schema.ts (Task 18)
The only consumer was `drizzle.config.ts`, which needs a TS file whose
top-level exports are the Drizzle table objects to introspect. Moving the
entry point next to the dialect bundles it consumes makes the package
layout match the responsibility: schema-related plumbing lives under
`src/schemas/`, services live under `src/services/`.

Replacement: `src/schemas/_internal/drizzle-kit-entry.ts`.

Behavior unchanged:
- Reads `DB_DIALECT` at module-load time and picks the matching dialect
  bundle from `schemas/_dialect-bundles/<dialect>`.
- Merges in the host app's `src/db/schemas/dynamic/index.ts` exports via
  `@nextly/scripts/merge-schemas` so user-defined tables show up in
  `drizzle-kit generate`.
- Same CommonJS-style top-level binding rewrite for drizzle-kit's static
  loader.

`drizzle.config.ts` now points at
`./src/schemas/_internal/drizzle-kit-entry.ts`. Underscore-prefix marks
the file as framework-internal — host apps should not import it
directly; they reach `getCoreSchema(dialect)` from `@nextly/schemas`
instead.
