+++ packages/nextly/drizzle.config.ts
@@ -12,7 +12,12 @@ if (!databaseUrl) {
 }
 
 export default defineConfig({
-  schema: "./src/services/lib/unified-schema.ts",
+  // Per Plan A Task 18: the prior schema entry point at
+  // `src/services/lib/unified-schema.ts` was deleted and moved to
+  // `src/schemas/_internal/drizzle-kit-entry.ts` alongside the dialect
+  // bundles it consumes. Same shape; same dialect-pick-at-load-time
+  // behavior; same drizzle-kit-visible top-level exports.
+  schema: "./src/schemas/_internal/drizzle-kit-entry.ts",
   out: `./src/database/migrations/${selectedDialect}`,
   dialect: selectedDialect,
   dbCredentials: { url: databaseUrl },
